Do not bail on aws-cli versions >= 2.0
Change version checking code that looks for minor version >= 8 to inspect
both major and minor versions and approve any version >= 1.8.

I'm not sure if the scripts all work with aws-cli 2.0, but I know that they
gave up without trying before, and this fixes that. It seems to run on my
system now.